28/May 2:00 PM

+++ app/Http/Controllers/Admin/AddressesController.php
@@ -8,6 +8,7 @@ use App\Models\CountryTranslation;
 use App\Models\Language;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\API\AddressesControllerApi;
 
 /**
  * Class AddressesController
@@ -21,26 +22,11 @@ class AddressesController extends Controller
      */
     public function getIndex()
     {
-        $countries = Country::all();
+        $address_api = new AddressesControllerApi();
 
-        // append translated country to all countries
-        foreach ($countries as $country) {
+        $getIndex_api = $address_api->getIndex();
 
-            // get country details
-            $country_translated = $country->translate();
-
-            // add the translated country as a key => value to main country object
-            // key is country_translated and the value id $country_details
-            $country->country_translated = $country_translated;
-
-            $city=City::where('country_id',$country->id)->first();
-
-            $city_translate=$city->translate();
-
-            $country->city_translate=$city_translate;
-        }
-
-        return view('admin.pages.addresses.index', compact('countries'));
+        return view('admin.pages.addresses.index', $getIndex_api['data']);
     }
 
 
@@ -49,6 +35,19 @@ class AddressesController extends Controller
      */
     public function getCreateNewAddress()
     {
+        $address_api = new AddressesControllerApi();
+
+        $getCreateNewAddress_api = $address_api->getCreateNewAddress();
+
+        if($getCreateNewAddress_api['status'] == false)
+        {
+            return [
+                'status' => 'error',
+                'title' => 'Error',
+                'text' => $getCreateNewAddress_api['msg']
+            ];
+        }
+
         return view('admin.pages.addresses.add-address');
     }
 
@@ -59,102 +58,24 @@ class AddressesController extends Controller
     public function createNewAddress(Request $request)
     {
 
-        // validation countries
-        $validation_countries = [
-            'country_name_en' => 'required',
-            'city_name_en' => 'required',
-        ];
+        $address_api = new AddressesControllerApi();
 
-        $validation = validator($request->all(), $validation_countries);
+        $createNewAddress_api = $address_api->createNewAddress($request);
 
-        // if validation failed, return false response
-        if ($validation->fails()) {
+        if($createNewAddress_api['status'] == false)
+        {
             return [
                 'status' => 'error',
-                'title' => $validation->getMessageBag(),
-                'text' => 'validation error'
+                'title' => ' Error',
+                'text' =>$createNewAddress_api['msg']
             ];
-        }
-
-        // choose one language to be the default one, let's make EN is the default
-        // store master country
-        // store the country in en
-
-        $en_id = Language::where('lang_code', 'en')->first()->id;
-        $ar_id = Language::where('lang_code', 'ar')->first()->id;
-
-
-        $country = Country::forceCreate([
-            'country_code' => $request->country_code,
-        ]);
-
-        // check saving success
-        if (!$country) {
-            return [
-                'status' => 'error',
-                'title' => 'Error',
-                'text' => 'something went wrong, please try again!'
-            ];
-        }
-
-        // store en version
-        $country_en = $country->countyTrans()->create([
-            'country' => $request->country_name_en,
-            'lang_id' => $en_id
-        ]);
-
-        // store ar version
-        // because it is not required, we check if there is ar in request, then save it, else {no problem, not required}
-        if ($request->country_name_ar) {
-
-            $country_ar = $country->countyTrans()->create([
-                'country' => $request->country_name_ar,
-                'lang_id' => $ar_id
-            ]);
-        }
-
-        // check if country save
-        if (!$country) {
-            //check save status
-            return [
-                'status' => 'error',
-                'title' => 'Error',
-                'msg' => 'something went wrong, please try again!'
-            ];
-        }
-
-        $country_id = $country->id;
-
-
-        foreach ($request->city_name_en as $key => $city_name_en) {
-
-            if ($city_name_en) {
-
-                // save master city
-                $city = City::forceCreate([
-                    'country_id' => $country_id,
-                ]);
-
-                $city_en = $city->cityTrans()->create([
-                    'city' => $city_name_en,
-                    'lang_id' => $en_id
-                ]);
-
-                if ($request->city_name_ar[$key]) {
-
-                    $city_ar = $city->cityTrans()->create([
-                        'city' => $request->city_name_ar[$key],
-                        'lang_id' => $ar_id
-                    ]);
-                }
-            }
         }
 
         // check success status
         return [
             'status' => 'success',
             'title' => 'successfully',
-            'text' => 'Data inserted successfully done',
+            'text' => $createNewAddress_api['msg']
         ];
     }
 
@@ -163,6 +84,19 @@ class AddressesController extends Controller
     */
     public function getAddCitiesTemplate()
     {
+        $address_api = new AddressesControllerApi();
+
+        $getAddCitiesTemplate_api = $address_api->getAddCitiesTemplate();
+
+        if($getAddCitiesTemplate_api['status'] == false)
+        {
+            return [
+                'status' => 'error',
+                'title' => 'Error',
+                'text' => $getAddCitiesTemplate_api['msg']
+            ];
+        }
+
         return view('admin.pages.addresses.templates.add-country.add-city')->render();
     }
 
@@ -171,6 +105,19 @@ class AddressesController extends Controller
     */
     public function getAddCitiesTemplateInEdit()
     {
+        $address_api = new AddressesControllerApi();
+
+        $getAddCitiesTemplateInEdit_api = $address_api->getAddCitiesTemplateInEdit();
+
+        if($getAddCitiesTemplateInEdit_api['status'] == false)
+        {
+            return [
+                'status' => 'error',
+                'title' => 'Error',
+                'text' => $getAddCitiesTemplateInEdit_api['msg']
+            ];
+        }
+
         return view('admin.pages.addresses.templates.add-country.add-city-in-edit')->render();
     }
 
@@ -181,23 +128,20 @@ class AddressesController extends Controller
      */
     public function getUpdateAddress($id)
     {
-        //get country by id
-        $country = Country::find($id);
-        //get country en
-        $country->en = $country->translate('en');
-        //get country ar
-        $country->ar = $country->translate('ar');
-        //get country id
+        $address_api = new AddressesControllerApi();
 
-        $country->cities;
+        $getUpdateAddress_api = $address_api->getUpdateAddress($id);
 
-        foreach ($country->cities as $city) {
-
-            $city->en = $city->translate('en');
-            $city->ar = $city->translate('ar');
+        if($getUpdateAddress_api['status'] == false)
+        {
+            return [
+                'status' => 'error',
+                'title' => 'Error',
+                'text' => $getUpdateAddress_api['msg']
+            ];
         }
 
-        return view('admin.pages.addresses.edit-address', compact('country'));
+        return view('admin.pages.addresses.edit-address', $getUpdateAddress_api['data']);
     }
 
     /**
@@ -207,125 +151,24 @@ class AddressesController extends Controller
      */
     public function updateAddress(Request $request)
     {
+        $address_api = new AddressesControllerApi();
 
-        // validation countries
-        $validation_countries = [
-            'country_code' => 'required',
-            'country_name_en' => 'required',
-            'cities_names_en' => 'required',
-        ];
+        $updateAddress_api = $address_api->updateAddress($request);
 
-        $validation = validator($request->all(), $validation_countries);
-
-        // if validation failed, return false response
-        if ($validation->fails()) {
-            return [
-                'status' => 'error',
-                'title' => $validation->getMessageBag(),
-                'text' => 'validation error'
-            ];
-        }
-
-        //find country by id
-        $country = Country::find($request->country_id);
-
-        //check if no country
-        if (!$country) {
+        if($updateAddress_api['status'] == false)
+        {
             return [
                 'status' => 'error',
                 'title' => 'Error',
-                'text' => 'There is no country with this id!'
+                'text' => $updateAddress_api['msg']
             ];
-        }
-
-        $country->country_code = $request->country_code;
-
-        if ($country->save()) {
-
-            $en_id = Language::where('lang_code', 'en')->first()->id;
-            $ar_id = Language::where('lang_code', 'ar')->first()->id;
-
-            $country->translate('en')->update([
-                'country' => $request->country_name_en
-            ]);
-
-            if ($request->country_name_ar) {
-
-                if ($country->translate('ar')) {
-
-                    $country->translate('ar')->update([
-                        'country' => $request->country_name_ar
-                    ]);
-
-                } else {
-
-                    $country->countyTrans()->create([
-                        'country' => $request->country_name_ar,
-                        'country_id' => $request->country_id,
-                        'lang_id' => $ar_id
-                    ]);
-                }
-            } else {
-
-                $country->translate('ar') ? $country->translate('ar')->delete() : '';
-            }
-
-
-            $cities_name_en = $request->cities_names_en;
-            $cities_name_ar = $request->cities_names_ar;
-
-            foreach ($cities_name_en as $key => $city_en) {
-
-                if (!$city_en) continue;
-
-                if (array_key_exists(1, $city_en)) {
-
-                    $city = City::find($city_en[1]);
-
-                    $city->cityTrans()->delete();
-
-                    $city->cityTrans()->create([
-                        'city' => $city_en[0],
-                        'lang_id' => $en_id,
-                        'city_id' => $city->id
-                    ]);
-
-                    if (array_key_exists($key, $cities_name_ar)) {
-
-                        $city->cityTrans()->create([
-                            'city' => $cities_name_ar[$key][0],
-                            'lang_id' => $ar_id,
-                            'city_id' => $city->id
-                        ]);
-                    }
-
-                    continue;
-                }
-
-                $city = City::forceCreate([
-                    'country_id' => $country->id,
-                ]);
-
-                $city->cityTrans()->create([
-                    'city' => $city_en[0],
-                    'lang_id' => $en_id
-                ]);
-
-                if (array_key_exists($key, $cities_name_ar)) {
-
-                    $city->cityTrans()->create([
-                        'city' => $cities_name_ar[$key][0],
-                        'lang_id' => $ar_id
-                    ]);
-                }
-            }
         }
 
         // check saving success
         return [
             'status' => 'success',
             'title', 'success',
-            'text' => 'Data updated successfully done',
+            'text' => $updateAddress_api['msg'],
         ];
     }
 
@@ -334,6 +177,19 @@ class AddressesController extends Controller
     */
     public function deleteCity($id)
     {
+        $address_api = new AddressesControllerApi();
+
+        $deleteCity_api = $address_api->deleteCity($id);
+
+        if($deleteCity_api['status'] == false)
+        {
+            return [
+                'status' => 'error',
+                'title' => 'Error',
+                'text' => $deleteCity_api['msg']
+            ];
+        }
+
         return City::find($id)->delete() ? ['status' => true] : ['status' => false];
     }
 
@@ -343,37 +199,24 @@ class AddressesController extends Controller
      */
     public function deleteAddress($id)
     {
-        //find country by id
-        $country = Country::find($id);
+        $address_api = new AddressesControllerApi();
 
-        // check if no city
-        if (!$country) {
+        $deleteAddress_api = $address_api->deleteAddress($id);
+
+        if($deleteAddress_api['status'] == false)
+        {
             return [
                 'status' => 'error',
                 'title' => 'Error',
-                'text' => 'There is no city!'
+                'text' => $deleteAddress_api['msg']
             ];
         }
-
-        $city = City::where('country_id', $id)->first();
-
-        // delete data from cityTrans
-        $city->cityTrans()->delete();
-
-        //delete data from city
-        $country->city()->delete();
-
-        //delete country
-        $country->countyTrans()->delete();
-
-        //delete data from country
-        $country->delete();
 
         //check save successfully
         return [
             'status' => 'success',
             'title' => 'success',
-            'text' => 'Data deleted successfully done!!'
+            'text' => $deleteAddress_api['msg']
         ];
     }
